+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added MIDI parser library.
 - Added MIDI Track data structure.
+- Added MIDI Note data structure.
 
 ### Fixed