+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added MIDI parser library.
 - Added MIDI Track data structure.
 - Added MIDI Note data structure.
+- Added MIDI file reader utility.
 
 ### Fixed