+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added *MIDI File Reader* node.
 - Added *MIDI Track Info* node.
 - Added *MIDI Note Info* node.
+- Added *Evaluate MIDI Track* node.
 
 ### Fixed